projeto/backends/django-backend/pyproject.toml
[project]
name = "django-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.5.0",
    "email-validator>=2.3.0",
    "gunicorn>=23.0.0",
    "openai>=2.8.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
